\name{NeuronIF-package}
\alias{NeuronIF-package}
\alias{NeuronIF}
\docType{package}
\title{
\packageTitle{NeuronIF}
}
\description{
\packageDescription{NeuronIF}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NeuronIF}
\packageIndices{NeuronIF}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{NeuronIF}

Maintainer: \packageMaintainer{NeuronIF}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
